renaming nexus/slack_connector to nexus/messaging_connector
