+++ justfile
@@ -102,19 +102,17 @@ setup-nexus:
 # Run the Slack connector worker. Safe to run multiple instances.
 # Requires: SLACK_BOT_TOKEN
 slack-connector:
-    cd "{{nexus_dir}}/slack_connector" && \
-    MESSAGING_PLATFORM=slack \
+    cd "{{nexus_dir}}/messaging_connector" && \
     SLACK_BOT_TOKEN="${SLACK_BOT_TOKEN}" \
     TEMPORAL_ADDRESS="${TEMPORAL_ADDRESS:-localhost:7233}" \
     CONNECTOR_NAMESPACE="${CONNECTOR_NAMESPACE:-connector}" \
     CONNECTOR_TASK_QUEUE="${CONNECTOR_TASK_QUEUE:-nexus-connector-slack}" \
-    go run ./cmd/worker/
+    go run ./cmd/slack/worker/
 
 # Run the Teams connector worker. Safe to run multiple instances.
 # Requires: MICROSOFT_TENANT_ID, MICROSOFT_APP_ID, MICROSOFT_APP_PASSWORD
 teams-connector:
-    cd "{{nexus_dir}}/slack_connector" && \
-    MESSAGING_PLATFORM=teams \
+    cd "{{nexus_dir}}/messaging_connector" && \
     MICROSOFT_TENANT_ID="${MICROSOFT_TENANT_ID}" \
     MICROSOFT_APP_ID="${MICROSOFT_APP_ID}" \
     MICROSOFT_APP_PASSWORD="${MICROSOFT_APP_PASSWORD}" \
@@ -122,32 +120,31 @@ teams-connector:
     TEMPORAL_ADDRESS="${TEMPORAL_ADDRESS:-localhost:7233}" \
     CONNECTOR_NAMESPACE="${CONNECTOR_NAMESPACE:-connector}" \
     CONNECTOR_TASK_QUEUE="${CONNECTOR_TASK_QUEUE:-nexus-connector-teams}" \
-    go run ./cmd/worker/
+    go run ./cmd/teams/worker/
 
 # Run the Slack webhook server.
 # Requires: SLACK_BOT_TOKEN, SLACK_SIGNING_SECRET
 slack-webhook:
-    cd "{{nexus_dir}}/slack_connector" && \
-    MESSAGING_PLATFORM=slack \
+    cd "{{nexus_dir}}/messaging_connector" && \
     SLACK_BOT_TOKEN="${SLACK_BOT_TOKEN}" \
     SLACK_SIGNING_SECRET="${SLACK_SIGNING_SECRET}" \
     TEMPORAL_ADDRESS="${TEMPORAL_ADDRESS:-localhost:7233}" \
     CONNECTOR_NAMESPACE="${CONNECTOR_NAMESPACE:-connector}" \
     CONNECTOR_TASK_QUEUE="${CONNECTOR_TASK_QUEUE:-nexus-connector-slack}" \
-    go run ./cmd/webhook/
+    go run ./cmd/slack/webhook/
 
 # Run the Teams webhook server.
 # Requires: MICROSOFT_TENANT_ID, MICROSOFT_APP_ID, MICROSOFT_APP_PASSWORD
 teams-webhook:
-    cd "{{nexus_dir}}/slack_connector" && \
-    MESSAGING_PLATFORM=teams \
+    cd "{{nexus_dir}}/messaging_connector" && \
     MICROSOFT_TENANT_ID="${MICROSOFT_TENANT_ID}" \
     MICROSOFT_APP_ID="${MICROSOFT_APP_ID}" \
     MICROSOFT_APP_PASSWORD="${MICROSOFT_APP_PASSWORD}" \
+    TEAMS_SERVICE_URL="${TEAMS_SERVICE_URL:-}" \
     TEMPORAL_ADDRESS="${TEMPORAL_ADDRESS:-localhost:7233}" \
     CONNECTOR_NAMESPACE="${CONNECTOR_NAMESPACE:-connector}" \
     CONNECTOR_TASK_QUEUE="${CONNECTOR_TASK_QUEUE:-nexus-connector-teams}" \
-    go run ./cmd/webhook/
+    go run ./cmd/teams/webhook/
 
 # Start a local Temporal dev server for the Monty example.
 temporal:
