+++ nexus/go.work
@@ -3,5 +3,5 @@ go 1.25.0
 use (
 	./agent_adapter/nexus_worker
 	./integration_tests
-	./slack_connector
+	./messaging_connector
 )
