+++ nexus/integration_tests/nexus_connector_durability_test.go
@@ -13,10 +13,10 @@ import (
 
 	"github.com/nexus-rpc/sdk-go/nexus"
 	"github.com/stretchr/testify/require"
-	agentiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/agent"
-	agentgen "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/agent/generated"
-	"github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/connector"
-	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging"
+	agentiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/agent"
+	agentgen "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/agent/generated"
+	"github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/connector"
+	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging"
 	commonpb "go.temporal.io/api/common/v1"
 	"go.temporal.io/sdk/activity"
 	"go.temporal.io/sdk/client"
