+++ nexus/messaging_connector/agent/contract_test.go
