+++ nexus/messaging_connector/agent/interfaces.go
