+++ nexus/messaging_connector/agent/temporal_native_harness_driver_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	agentgen "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/agent/generated"
-	ncmsg "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging"
+	agentgen "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/agent/generated"
+	ncmsg "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging"
 	commonpb "go.temporal.io/api/common/v1"
 	"go.temporal.io/sdk/activity"
 	"go.temporal.io/sdk/testsuite"
