+++ nexus/messaging_connector/messaging/slack/messaging_driver_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	slackapi "github.com/slack-go/slack"
-	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging"
+	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
