+++ nexus/messaging_connector/messaging/teams/webhook/server.go
@@ -8,10 +8,10 @@ import (
 	"net/http"
 	"strings"
 
-	agentiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/agent"
-	"github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/connector"
-	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging"
-	"github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging/teams"
+	agentiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/agent"
+	"github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/connector"
+	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging"
+	"github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging/teams"
 	"go.temporal.io/sdk/client"
 )
