+++ nexus/messaging_connector/messaging/teams/webhook/server_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/slack_connector/messaging"
+	msgiface "github.com/temporal-community/temporal-agent-harness/nexus/messaging_connector/messaging"
 )
 
 func TestMessageWorkflowInputPropagatesConversationType(t *testing.T) {
