+++ nexus/messaging_connector/slack_app_manifest.yaml
